+++ NEWS.md
@@ -1,8 +1,9 @@
 # tbgeneratr 0.0.0.21.9003
-#### 16 Jan 2019
+#### 17 Jan 2019
 ## Major changes
 * `aggregate_dst()` added to simplify `dst_baseliner()`.
 * `dst_baseliner.epiinfo()` added to package.
+* `dst_baseliner.koch6()` added to package - option for Gr lab data only.
 
 ## Minor changes
 * `drug_baseliner()` args changed.